requirements.txt
# FNX Talent Factory - Python Dependencies

# LLM Providers
google-genai>=1.0.0       # Google Gemini API (primary)
openai>=1.0.0             # OpenAI API (fallback)

# API Server
fastapi>=0.104.0          # Web framework
uvicorn>=0.24.0           # ASGI server
python-dotenv>=1.0.0      # .env file loading

# Phase 2 (uncomment when ready):
# supabase>=2.0.0         # Supabase Python client
